+++ src/interrupt.rs
@@ -196,12 +196,21 @@ fn exit(_cs: &IsrCriticalSection) {
     }
 }
 
-#[cfg(any(esp32, esp32s2, esp32s3, esp32p4))]
+#[cfg(any(esp32, esp32s2, esp32s3))]
 #[inline(always)]
 #[link_section = ".iram1.interrupt_exit"]
 fn exit(cs: &IsrCriticalSection) {
     unsafe {
         vPortExitCritical(cs.0.get());
+    }
+}
+
+#[cfg(esp32p4)]
+#[inline(always)]
+#[link_section = ".iram1.interrupt_exit"]
+fn exit(cs: &IsrCriticalSection) {
+    unsafe {
+        vPortExitCriticalMultiCore(cs.0.get());
     }
 }