+++ src/rmt/tx_channel.rs
@@ -81,6 +81,7 @@ impl<'d> TxChannelDriver<'d> {
                     },
                     #[cfg(any(
                         esp_idf_version_patch_at_least_5_4_3,
+                        esp_idf_version_at_least_5_5_2,
                         esp_idf_version_at_least_6_0_0
                     ))]
                     {